# perf: [date-2.parse] Add a fastpath for timestamps that are likely ISO timestamps