In a synthetic benchmark I've made for myself (Posgres, a table of soccer matches with 100k rows), rendering a card takes ~1.6 sec on average (as reported by "How long the query took" status). The table contains mostly plain values and one column of timestamps (dates). A large chunk of that time is spent parsing Strings into datetimes with `metabase.util.date-2.parse`. That function uses a complicated composite parser that can handle many different timestamp representations. However, my hunch (correct me here if I'm wrong) is that most databases would yield timestamps in a relatively stable format, like Postgres returns an ISO_INSTANT timestamp `"1970-01-01T00:00:00Z"`.

The idea is to check if the timestamp has length that matches ISO_INSTANT format, and if so, try to parse it with the much faster DateTimeFormatter.ISO_INSTANT parser. If parsing fails, fall back to the regular composite parser. If the fast path is successful, we get a tremendous improvement of parsing a single datetime in ~800 ns (compared to ~4.4 us for the full parser). The worst drawback is adding ~1us for failed attempts per timestamp that is exactly 20 characters long but is not an INSTANT.

After the patch, the card rendering time improves twofold: from 1.6 sec to 800 ms.