@alexander-yakushev 2024-08-22
If there are other common timestamp formats that some databases return dates in, it might make sense to add fastpathes for those too. I defer to others' knowledge on this.